added $(dir)_SCRIPTS target for installing scripts

+++ src/exe.mk
@@ -11,6 +11,8 @@
 # the skeleton should set the following variables
 #    dir must be set to the subdirectory name (relative to the location of Makefile)
 #    $(dir)_SOURCES must be set to a list of .cxx and/or .c files (without path)
+#    $(dir)_SCRIPTS must be set to a list of files (presumably shell scripts) that need
+#      to be installed
 # Result:
 # targets $(dir) clean_exes_$(dir) install_exes_$(dir), uninstall_exes_$(dir)
 #    and for each file in $(dir)_SOURCES there will be a (phony) target
@@ -23,17 +25,18 @@
 #
 # make SUBDIR              -> will compile and link all executables
 # make clean_exes_SUBDIR   -> will remove all generated files
-# make install_exes_SUBDIR      -> will install all executables
+# make install_exes_SUBDIR -> will install all executables
 # make $(dir)/file         -> will compile and link single file
 #                             (there is no need to use the $(DEST) prefix)
 
 $(dir)_SOURCES:= $(addprefix $(dir)/, $($(dir)_SOURCES))
 $(dir)_EXES:= \
-	$(patsubst %.cxx, $(DEST)%, $(filter %.cxx, $($(dir)_SOURCES))) \
-	$(patsubst %.c, $(DEST)%, $(filter %.c, $($(dir)_SOURCES)))
+        $(patsubst %.cxx, $(DEST)%, $(filter %.cxx, $($(dir)_SOURCES))) \
+        $(patsubst %.c, $(DEST)%, $(filter %.c, $($(dir)_SOURCES)))
 $(dir)_EXES_without_DEST:= \
-	$(patsubst %.cxx, %, $(filter %.cxx, $($(dir)_SOURCES))) \
-	$(patsubst %.c, %, $(filter %.c, $($(dir)_SOURCES)))
+        $(patsubst %.cxx, %, $(filter %.cxx, $($(dir)_SOURCES))) \
+        $(patsubst %.c, %, $(filter %.c, $($(dir)_SOURCES)))
+$(dir)_SCRIPTS:= $(addprefix $(dir)/, $($(dir)_SCRIPTS))
 
 .PHONY: $(dir) clean_exes_$(dir) install_exes_$(dir) uninstall_exes_$(dir)
 
@@ -63,25 +66,27 @@ $(dir)_EXE_FILENAMES := $(addsuffix $(EXE_SUFFIX), $($(dir)_EXES))
 # note: see lib.mk for explanation for the $(@:...) trick
 # it really is just a way to get $(dir) at build-time
 clean_exes_$(dir):
-	rm -f $($(@:clean_exes_%=%)_EXE_FILENAMES)
-	rm -f $(DEST)$(@:clean_exes_%=%)/*.[P]
-	rm -f $(DEST)$(@:clean_exes_%=%)/*$(O_SUFFIX)
+        rm -f $($(@:clean_exes_%=%)_EXE_FILENAMES)
+        rm -f $(DEST)$(@:clean_exes_%=%)/*.[P]
+        rm -f $(DEST)$(@:clean_exes_%=%)/*$(O_SUFFIX)
 
 
 
 install_exes_$(dir): $(dir)
-	mkdir -p $(INSTALL_EXE_DIR)
-	if [ ! -z "$($(@:install_exes_%=%)_EXE_FILENAMES)" ]; then $(INSTALL) $($(@:install_exes_%=%)_EXE_FILENAMES) $(INSTALL_EXE_DIR); fi
+        mkdir -p $(INSTALL_EXE_DIR)
+        if [ ! -z "$($(@:install_exes_%=%)_EXE_FILENAMES)" ]; then $(INSTALL) $($(@:install_exes_%=%)_EXE_FILENAMES) $(INSTALL_EXE_DIR); fi
+        if [ ! -z "$($(@:install_exes_%=%)_SCRIPTS)" ]; then $(INSTALL) $($(@:install_exes_%=%)_SCRIPTS) $(INSTALL_EXE_DIR); fi
 
 uninstall_exes_$(dir):
-	$(RM) $(addprefix $(INSTALL_EXE_DIR)/, $(notdir $($(@:uninstall_exes_%=%)_EXE_FILENAMES)))
+        if [ ! -z "$($(@:install_exes_%=%)_EXE_FILENAMES)" ]; then $(RM) $(addprefix $(INSTALL_EXE_DIR)/, $(notdir $($(@:uninstall_exes_%=%)_EXE_FILENAMES))); fi
+        if [ ! -z "$($(@:install_exes_%=%)_SCRIPTS)" ]; then $(RM) $(addprefix $(INSTALL_EXE_DIR)/, $(notdir $($(@:uninstall_exes_%=%)_SCRIPTS))); fi
 
 install_$(dir): install_exes_$(dir)
 
 ifneq ($(MAKECMDGOALS:clean%=clean),clean)
   -include \
-	$(patsubst %.cxx, $(DEST)%.P, $(filter %.cxx, $($(dir)_SOURCES))) \
-	$(patsubst %.c, $(DEST)%.P, $(filter %.c, $($(dir)_SOURCES)))
+        $(patsubst %.cxx, $(DEST)%.P, $(filter %.cxx, $($(dir)_SOURCES))) \
+        $(patsubst %.c, $(DEST)%.P, $(filter %.c, $($(dir)_SOURCES)))
 endif
 
 # set to some garbage such that we get an error when  the next skeleton forgets to set dir
